+++ src/Drivers/DPDK/DpdkDriverImpl.h
@@ -1,4 +1,4 @@
-/* Copyright (c) 2019, Stanford University
+/* Copyright (c) 2019-2020, Stanford University
  *
  * Permission to use, copy, modify, and/or distribute this software for any
  * purpose with or without fee is hereby granted, provided that the above
@@ -110,7 +110,7 @@ class DpdkDriver::Impl {
         explicit Packet(OverflowBuffer* overflowBuf);
 
         /// see Driver::Packet::getMaxPayloadSize()
-        virtual uint16_t getMaxPayloadSize()
+        virtual int getMaxPayloadSize()
         {
             return MAX_PAYLOAD_SIZE;
         }
