+++ src/Drivers/Fake/FakeDriver.h
@@ -1,4 +1,4 @@
-/* Copyright (c) 2019, Stanford University
+/* Copyright (c) 2019-2020, Stanford University
  *
  * Permission to use, copy, modify, and/or distribute this software for any
  * purpose with or without fee is hereby granted, provided that the above
@@ -16,13 +16,13 @@
 #ifndef HOMA_DRIVERS_FAKE_FAKEDRIVER_H
 #define HOMA_DRIVERS_FAKE_FAKEDRIVER_H
 
+#include <Homa/Driver.h>
+#include <Homa/Drivers/Util/QueueEstimator.h>
+
 #include <array>
 #include <chrono>
 #include <deque>
 #include <mutex>
-
-#include <Homa/Driver.h>
-#include <Homa/Drivers/Util/QueueEstimator.h>
 
 namespace Homa {
 namespace Drivers {
@@ -77,7 +77,7 @@ class FakePacket : public Driver::Packet {
     virtual ~FakePacket() {}
 
     /// see Driver::Packet::getMaxPayloadSize()
-    virtual uint16_t getMaxPayloadSize()
+    virtual int getMaxPayloadSize()
     {
         return MAX_PAYLOAD_SIZE;
     }
